339.0.0 release

+++ google-cloud-sdk/lib/googlecloudsdk/command_lib/storage/tasks/cp/file_part_upload_task.py
@@ -32,6 +32,7 @@ from googlecloudsdk.command_lib.storage import errors
 from googlecloudsdk.command_lib.storage import file_part
 from googlecloudsdk.command_lib.storage import hash_util
 from googlecloudsdk.command_lib.storage import progress_callbacks
+from googlecloudsdk.command_lib.storage import storage_url
 from googlecloudsdk.command_lib.storage.tasks import task_status
 from googlecloudsdk.command_lib.storage.tasks.cp import file_part_task
 from googlecloudsdk.command_lib.storage.tasks.rm import delete_object_task
@@ -77,7 +78,14 @@ class FilePartUploadTask(file_part_task.FilePartTask):
         self._source_resource.storage_url.object_name)
     provider = self._destination_resource.storage_url.scheme
 
-    digesters = {hash_util.HashAlgorithm.MD5: hash_util.get_md5()}
+    # Do not compute hashes if the user provided an md5, or if the upload uses
+    # boto3, which performs its own unskippable validation.
+    if (self._source_resource.md5_hash or
+        provider == storage_url.ProviderPrefix.S3):
+      digesters = {}
+    else:
+      digesters = {hash_util.HashAlgorithm.MD5: hash_util.get_md5()}
+
     with file_part.FilePart(source_stream, self._offset,
                             self._length, digesters=digesters) as upload_stream:
       destination_resource = api_factory.get_api(provider).upload_object(
@@ -87,15 +95,15 @@ class FilePartUploadTask(file_part_task.FilePartTask):
               md5_hash=self._source_resource.md5_hash, size=self._length),
           progress_callback=progress_callback)
 
-    # TODO(b/175904829): Skip this if the hash is provided through a flag.
-    calculated_digest = hash_util.get_base64_hash_digest_string(
-        digesters[hash_util.HashAlgorithm.MD5])
-    try:
-      hash_util.validate_object_hashes_match(self._source_resource.storage_url,
-                                             calculated_digest,
-                                             destination_resource.md5_hash)
-    except errors.HashMismatchError:
-      delete_object_task.DeleteObjectTask(
-          destination_resource.storage_url).execute(
-              task_status_queue=task_status_queue)
-      raise
+    if digesters:
+      calculated_digest = hash_util.get_base64_hash_digest_string(
+          digesters[hash_util.HashAlgorithm.MD5])
+      try:
+        hash_util.validate_object_hashes_match(
+            self._source_resource.storage_url, calculated_digest,
+            destination_resource.md5_hash)
+      except errors.HashMismatchError:
+        delete_object_task.DeleteObjectTask(
+            destination_resource.storage_url).execute(
+                task_status_queue=task_status_queue)
+        raise
